FIX #18863 TIME 0:02 fix somes mistakes in code

+++ src/app/entity/controllers/EntityController.php
@@ -541,7 +541,9 @@ class EntityController
         ]);
         //ResourceContact
         $dyingConnection = ResourceContactModel::get(['select' => ['id', 'res_id', 'item_id', 'mode'], 'where' => ['type = ?', 'item_id = ?'], 'data' => ['entity', $dyingEntity['id']]]);
-        $successorConnection = ResourceContactModel::get(['select' => ['id', 'res_id', 'item_id', 'mode'], 'where' => ['type = ?', 'item_id = ?', 'res_id in (?)'], 'data' => ['entity', $successorEntity['id'], array_uniq(array_column($dyingConnection, 'res_id'))]]);
+        if(!empty($dyingConnection)) {
+            $successorConnection = ResourceContactModel::get(['select' => ['id', 'res_id', 'item_id', 'mode'], 'where' => ['type = ?', 'item_id = ?', 'res_id in (?)'], 'data' => ['entity', $successorEntity['id'], array_uniq(array_column($dyingConnection, 'res_id'))]]);
+        }
         $dyingIds = array_column($dyingConnection, 'id');
         $idsToDelete = [];
         foreach ($dyingConnection as $dyingConn) {
@@ -551,9 +553,14 @@ class EntityController
                 }
             }
         }
-        ResourceContactModel::delete(['where' => ['id in (?)'], 'data' => [$idsToDelete]]);
-        ResourceContactModel::update(['set' => ['item_id' => $successorEntity['id']], 'where' => ['id in (?)'], 'data' => [$dyingIds]]);
-
+        if(!empty($idsToDelete)) {
+            ResourceContactModel::delete(['where' => ['id in (?)'], 'data' => [$idsToDelete]]);
+        }
+        if(!empty($dyingIds)) {
+            ResourceContactModel::update(['set' => ['item_id' => $successorEntity['id']], 'where' => ['id in (?)'], 'data' => [$dyingIds]]);
+        }
+        EntityModel::delete(['where' => ['entity_id = ?'], 'data' => [$aArgs['id']]]);
+        
         HistoryController::add([
             'tableName' => 'entities',
             'recordId'  => $aArgs['id'],
